requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via starlette
click==8.1.7
    # via uvicorn
dnspython==2.6.1
    # via email-validator
email-validator==2.1.1
    # via -r requirements.in
fastapi==0.110.1
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.6
    # via
    #   anyio
    #   email-validator
jinja2==3.1.3
    # via -r requirements.in
loguru==0.7.2
    # via -r requirements.in
markupsafe==2.1.5
    # via jinja2
packaging==24.0
    # via gunicorn
pydantic==2.6.4
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.16.3
    # via pydantic
pyyaml==6.0.1
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.37.2
    # via fastapi
typing-extensions==4.11.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn==0.29.0
    # via -r requirements.in
watchdog==4.0.0
    # via -r requirements.in
